+++ app/src/main/java/com/example/kolin/testapplication/domain/Food.java
@@ -8,7 +8,7 @@ public class Food {
 
     private Double b;
     private String idName;
-    private Integer j;
+    private Double j;
     private String name;
     private Integer weight;
     private Double y;
@@ -29,11 +29,11 @@ public class Food {
         this.idName = idName;
     }
 
-    public Integer getJ() {
+    public Double getJ() {
         return j;
     }
 
-    public void setJ(Integer j) {
+    public void setJ(Double j) {
         this.j = j;
     }
